Move all Route handlers into web/pages
Solves #433
